plugins/systemstats/pyproject.toml
[project]
name = "coremind-plugin-systemstats"
version = "0.1.0"
description = "System statistics sensor plugin for CoreMind."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Guillaume Gagnon" }]
requires-python = ">=3.12"
keywords = ["coremind", "plugin", "system", "statistics"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "coremind>=0.0.1",
    "grpcio>=1.66.0",
    "psutil>=6.0.0",
    "structlog>=24.4.0",
]

[project.scripts]
coremind-plugin-systemstats = "coremind_plugin_systemstats.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coremind_plugin_systemstats"]
